+++ services/ecs/v2/EcsClient.ts
@@ -959,7 +959,7 @@ export class EcsClient {
      * @param {string} [enterpriseProjectId] 查询绑定某个企业项目的弹性云服务器。  若需要查询当前用户所有企业项目绑定的弹性云服务，请传参all_granted_eps。
      * @param {string} [flavor] 云服务器规格ID,已上线的规格请参见《弹性云服务器用户指南》的“实例和应用场景”章节。
      * @param {string} [ip] IPv4地址过滤结果，匹配规则为模糊匹配。
-     * @param {number} [limit] 查询返回云服务器当前页面的大小。每页最多返回1000台云服务器的信息。
+     * @param {number} [limit] 查询返回云服务器当前页面的大小。每页默认值是25，最多返回1000台云服务器的信息。
      * @param {string} [name] 云服务器名称，匹配规则为模糊匹配。
      * @param {string} [notTags] 查询tag字段中不包含该值的云服务器。
      * @param {number} [offset] 页码。 当前页面数，默认为1。  取值大于等于0，取值为0时返回第1页。