+++ services/image/v2/model/ImageDescriptionReq.ts
@@ -3,7 +3,6 @@
 export class ImageDescriptionReq {
     public image?: string;
     public url?: string;
-    public language?: string;
     public constructor() { 
     }
     public withImage(image: string): ImageDescriptionReq {
@@ -12,10 +11,6 @@ export class ImageDescriptionReq {
     }
     public withUrl(url: string): ImageDescriptionReq {
         this['url'] = url;
-        return this;
-    }
-    public withLanguage(language: string): ImageDescriptionReq {
-        this['language'] = language;
         return this;
     }
 }